+++ src/main/java/Main.java
@@ -1,78 +1,5 @@
-import controlers.ControlerSimulation;
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.*;
-import javafx.stage.Stage;
-import controlers.ControlerVue;
-import modele.*;
-import views.*;
-
-public class Main extends Application {
-
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        //Creation du projet (le modèle)
-        Modele modele = new Modele();
-
-        //Contrôleur pour choisir la vue
-        ControlerVue controlVue = new ControlerVue(modele);
-        //Controler de l'affichage de la simulation
-        ControlerSimulation controlerSimulation = new ControlerSimulation(modele);
-
-        //Les Vues
-        VueSimulationMenu vueSimulationMenu = new VueSimulationMenu();
-        VueSimulationCreate vueSimulationCreate = new VueSimulationCreate();
-        VueSimulationChoixPartie vueSimulationChoixPartie = new VueSimulationChoixPartie();
-        VueSimulationMain vueSimulationMain = new VueSimulationMain();
-        VueApprentissageMenu vueApprentissageMenu = new VueApprentissageMenu();
-        VueApprentissageMain vueApprentissageMain = new VueApprentissageMain();
-        VueCartes vueCartes = new VueCartes(10);
-        //Les vues s'enregistrent comme vues du modele
-        modele.enregistrerObservateur(vueSimulationMenu);
-        modele.enregistrerObservateur(vueSimulationCreate);
-        modele.enregistrerObservateur(vueSimulationChoixPartie);
-        modele.enregistrerObservateur(vueSimulationMain);
-        modele.enregistrerObservateur(vueApprentissageMenu);
-        modele.enregistrerObservateur(vueApprentissageMain);
-        modele.enregistrerObservateur(vueCartes);
-
-        //La page principale
-        BorderPane borderPane = new BorderPane();
-
-        //la partie avec les boutons pour accéder aux autres vues
-        HBox head = new HBox();
-        //les boutons pour accéder aux vues
-        Button buttonVueSimulation = new Button("Simulation");
-        Button buttonVueApprentissage = new Button("Apprentissage");
-        Button buttonVueCartes = new Button("Cartes");
-        Button buttonQuitter = new Button("Quitter");
-        //ajout du controle des boutons à controlerVue
-        buttonVueSimulation.setOnMouseClicked(controlVue);
-        buttonVueApprentissage.setOnMouseClicked(controlVue);
-        buttonVueCartes.setOnMouseClicked(controlVue);
-        buttonQuitter.setOnMouseClicked(controlVue);
-        //on ajoute les boutons à la Page
-        head.getChildren().addAll(buttonVueSimulation, buttonVueApprentissage, buttonVueCartes, buttonQuitter);
-        borderPane.setTop(head);
-
-        // Centre avec les vues
-        VBox centerBox = new VBox(vueSimulationMenu, vueSimulationCreate, vueSimulationChoixPartie, vueSimulationMain, vueApprentissageMenu, vueApprentissageMain, vueCartes);
-        borderPane.setCenter(centerBox);
-
-        //scene et stage
-        Scene scene = new Scene(borderPane,1200,600);
-        stage.setScene(scene);
-        stage.setTitle("CTF");
-        stage.show();
-
-        //on actualise pour tout afficher
-        modele.notifierObservateurs();
+public class Main {
+    public static void main(final String[] args) {
+        App.main(args);
     }
 }
