+++ src/main/java/display/DisplaySimulation.java
@@ -3,10 +3,7 @@ package display;
 import engine.Coordinate;
 import engine.Team;
 import engine.agent.Agent;
-import engine.map.Cell;
-import engine.map.GameMap;
-import engine.map.Ground;
-import engine.map.Wall;
+import engine.map.*;
 import engine.object.Flag;
 import engine.object.GameObject;
 import javafx.geometry.Insets;
@@ -24,9 +21,9 @@ import java.util.List;
 
 //Display qui affiche la carte de jeu avec les boutons
 public class DisplaySimulation extends Display {
-    int tailleCase = 32;
+    int tailleCase = 64;
 
-    public DisplaySimulation(Node simulationBox) {
+    public DisplaySimulation(Pane simulationBox) {
         super(simulationBox);
     }
 
@@ -38,13 +35,24 @@ public class DisplaySimulation extends Display {
         for (int ligne = 0; ligne < cells.size(); ligne++) {
             for (int colonne = 0; colonne < cells.get(ligne).size(); colonne++) {
                 Cell cell = cells.get(ligne).get(colonne);
-                Image sprite = null;
-                if (cell instanceof Ground) {
-                    sprite = new Image("file:ressources/top/sol_neutre.png", tailleCase, tailleCase, false, false);
+                String pathImageCell = "file:ressources/top/";
+                //Détection du type de case
+                if (cell instanceof Ground || cell instanceof SpawningCell) {
+                    //Détection de l'équipe
+                    if (cell.getTeam() == Team.BLUE) {
+                        pathImageCell +="sol_vert.png";
+                    }
+                    else if (cell.getTeam() == Team.PINK) {
+                        pathImageCell +="sol_rouge.png";
+                    }
+                    else {
+                        pathImageCell +="sol_neutre.png";
+                   }
                 }
                 else if (cell instanceof Wall) {
-                    sprite = new Image("file:ressources/top/mur_vue_haut.png", tailleCase, tailleCase, false, false);
+                    pathImageCell += "mur_vue_haut.png";
                 }
+                Image sprite = new Image(pathImageCell, tailleCase, tailleCase, false, false);
                 ImageView imageView = new ImageView(sprite);
                 GridPane.setConstraints(imageView, colonne, ligne);
                 grilleMap.getChildren().add(imageView);
@@ -53,7 +61,7 @@ public class DisplaySimulation extends Display {
         //Stack Pane pour stocker la carte + Les objets dessus (agents)
         StackPane stackPane = new StackPane(grilleMap);
         for (Agent agent : agents) {
-            double tailleAgent = agent.getRadius();
+            double tailleAgent = 32;
             String pathImageAgent = "file:ressources/top/robot/";
             if (agent.getTeam() == Team.BLUE) {
                 pathImageAgent += "Rouge/robot_rose_flat_haut.png";
@@ -63,23 +71,26 @@ public class DisplaySimulation extends Display {
             }
             Image spriteAgent = new Image(pathImageAgent, tailleAgent, tailleAgent, false, false);
             ImageView agentView = new ImageView(spriteAgent);
-            agentView.setX(agent.getCoordinate().x());
-            agentView.setY(agent.getCoordinate().y());
+            agentView.setTranslateX(agent.getCoordinate().x());
+            agentView.setTranslateY(agent.getCoordinate().y());
             stackPane.getChildren().add(agentView);
         }
         for (GameObject object : objects) {
             String pathImageObjet = "file:ressources/top/";
             if (object instanceof Flag) {
-                pathImageObjet += "drapeau_take_bleu_left.png";
+                if (((Flag) object).getTeam() == Team.PINK) {
+                    pathImageObjet += "drapeau_rouge.png";
+                }
+                else {
+                    pathImageObjet += "drapeau_bleu.png";
+                }
             }
             Image spriteAgent = new Image(pathImageObjet, tailleCase, tailleCase, false, false);
             ImageView agentView = new ImageView(spriteAgent);
-            agentView.setX(object.getCoordinate().x());
-            agentView.setY(object.getCoordinate().y());
+            agentView.setTranslateX(object.getCoordinate().x());
+            agentView.setTranslateY(object.getCoordinate().y());
             stackPane.getChildren().add(agentView);
         }
-
-        //System.out.println("bonjour");
 
         //Le display est uniquement le stackpane
         root.getChildren().clear();
