+++ src/main/java/engine/object/Flag.java
@@ -11,4 +11,11 @@ public class Flag extends GameObject {
         super(coordinate);
         this.team = team;
     }
+
+    public void setTeam(Team team) {
+        this.team = team;
+    }
+    public Team getTeam() {
+        return team;
+    }
 }
