+++ src/main/java/views/Observateur.java
@@ -6,5 +6,5 @@ import modele.Modele;
  * Interface pour les differents observateurs de MVC.Sujet
  */
 public interface Observateur {
-	public void actualiser(Modele modele);
+	public void actualiser(Modele modele) throws Exception;
 }
